requirements.txt
asgiref==3.4.1
gunicorn==20.1.0
Django==3.2.7
django-cors-headers==3.9.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
numpy==1.21.2
pandas==1.3.4
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.3
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
whitenoise==5.3.0
